requirements.txt
fastapi[all]
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-experimental
pypdfium2
tiktoken
open-clip-torch 
tqdm
pdfminer-six
python-docx
loguru
chromadb
Unidecode
langserve[all]
pillow
torch

